localstorage and broke down to smaller components
